pyproject.toml
[project]
name = "basis-expert-council"
version = "0.1.0"
description = "BasisPilot 贝领 — Your AI Co-Pilot Through BASIS · 基于 DeepAgents/LangGraph 架构"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "xDAN-AI" },
]
keywords = ["education", "BASIS", "BasisPilot", "AI", "agent", "LangGraph", "tutoring", "AP"]

dependencies = [
    "deepagents>=0.1.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langgraph>=0.4.0",
    "python-dotenv>=1.0.0",
    "asyncpg>=0.30.0",
    "httpx>=0.27.0",
    "PyJWT>=2.9.0",
    "mem0ai>=0.1.0",
    "pydantic>=2.0.0",
    "PyYAML>=6.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.8.0",
]

[project.scripts]
basis-expert = "basis_expert_council.agent:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/basis_expert_council"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
